Isolate Hub state integration workloads
Preserve loadfile scheduling and all 42 HubMgmt cases while separating plain-Hub dataplane fixtures from routing dependencies. Exercise both fresh migration destinations and leave final dataplane cleanup to the checked fixture.
